+++ src/main/java/com/lawschool/service/impl/CollectionServiceImpl.java
@@ -54,7 +54,7 @@ public class CollectionServiceImpl implements CollectionService {
             collection.setComUserid(user.getId());
             collection.setOptuser(user.getFullName());
             collection.setOpttime(new Date());
-            collection.setDelStatus((short)0);
+            collection.setDelStatus(0);
             return collectionDao.insert(collection)==1?SUCCESS:ERROR;//1 添加成功  0 添加失败
         }
         if(existStatus[0].equals(ONCE_EXIST+"")){//重新收藏
@@ -86,7 +86,7 @@ public class CollectionServiceImpl implements CollectionService {
         EntityWrapper<Collection> ew=new EntityWrapper();
         ew.eq("COM_USERID",user.getId()).eq("ID",collection.getId());
         if(flag){//0-》1  取消收藏
-            collection.setDelStatus((short)1);
+            collection.setDelStatus(1);
             ew.eq("DEL_STATUS",0);
         }else {//1-》0 重新收藏
             collection.setDelStatus((short)0);
